Rename ResultMessageFactory classes
